wlineedit - added two convenience functions to set output & input styles
Changed getOutput/Input style functions to return areference to const
QString class members, "const QString&", to correctly match what's
expected by QLineEdit::setStyleSheet(const QString&). Added QString
members to the private section of the class to facilitate this change.

Set alignment of the edit box to left when it gets the focus, so that
input is entered from left to right. This change was made because users
of this class have the option to set the alignment of the text in the
box to left or right, but we want input to go from the left to the
right.

Added comments.